`TrainerAbstract` and related updates for `TrainerTTS`
